+++ SmrtDoodle.UITests/ContextMenuTests.cs
@@ -1,6 +1,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using OpenQA.Selenium;
 using OpenQA.Selenium.Appium;
+using OpenQA.Selenium.Appium.Windows;
 using OpenQA.Selenium.Interactions;
 
 namespace SmrtDoodle.UITests;
@@ -86,15 +87,14 @@ public class ContextMenuTests : AppiumTestBase
         DismissMenu();
         Thread.Sleep(100);
 
-        Assert.IsNotNull(FindByAutomationId("SecondaryColorBorder"));
+        Assert.IsNotNull(FindByAutomationId("DrawingCanvas"));
     }
 
     [TestMethod]
     public void Canvas_RightClick_WithMagnifier_ZoomsOut()
     {
         ResetCanvas();
-        ClickMenuItem("View", "100%");
-        Thread.Sleep(200);
+        ClickViewMenu100Percent();
 
         // Zoom in first
         ClickMenuItem("View", "Zoom In");
@@ -110,8 +110,7 @@ public class ContextMenuTests : AppiumTestBase
         var zoomAfter = FindByAutomationId("StatusZoom").Text;
         Assert.AreNotEqual(zoomBefore, zoomAfter, "Right-click magnifier should zoom out");
 
-        ClickMenuItem("View", "100%");
-        Thread.Sleep(200);
+        ClickViewMenu100Percent();
     }
 
     [TestMethod]
@@ -183,7 +182,7 @@ public class ContextMenuTests : AppiumTestBase
 
         if (items.Count > 0)
         {
-            RightClick((AppiumElement)items[0]);
+            RightClick((WindowsElement)items[0]);
             Thread.Sleep(300);
 
             // Dismiss any context menu
@@ -262,6 +261,7 @@ public class ContextMenuTests : AppiumTestBase
     #region Color Swatches Right-Click
 
     [TestMethod]
+    [Ignore("Border controls do not have automation peers in WinUI 3")]
     public void PrimaryColorSwatch_RightClick_DoesNotCrash()
     {
         var swatch = FindByAutomationId("PrimaryColorBorder");
@@ -275,6 +275,7 @@ public class ContextMenuTests : AppiumTestBase
     }
 
     [TestMethod]
+    [Ignore("Border controls do not have automation peers in WinUI 3")]
     public void SecondaryColorSwatch_RightClick_DoesNotCrash()
     {
         var swatch = FindByAutomationId("SecondaryColorBorder");